Move ProvinceDefinitions from Mappers to HoI4 world
