+++ Vic2ToHoI4/Source/HOI4World/HoI4World.cpp
@@ -1,8 +1,8 @@
 #include "HoI4World.h"
 #include "../Configuration.h"
+#include "../HOI4World/ProvinceDefinitions.h"
 #include "../Mappers/CountryMapping.h"
 #include "../Mappers/FlagsToIdeas/FlagsToIdeasMapper.h"
-#include "../Mappers/ProvinceDefinitions.h"
 #include "../Mappers/TechMapper.h"
 #include "../V2World/Agreement.h"
 #include "../V2World/Country.h"
