+++ Vic2ToHoI4/Source/HOI4World/Map/MapData.cpp
@@ -1,7 +1,7 @@
 #include "MapData.h"
 #include "../../Color.h"
 #include "../../Configuration.h"
-#include "../../Mappers/ProvinceDefinitions.h"
+#include "../../HOI4World/ProvinceDefinitions.h"
 #include "Log.h"
