+++ Vic2ToHoI4/Source/HOI4World/Navies/Navies.cpp
@@ -1,5 +1,5 @@
 #include "Navies.h"
-#include "../../Mappers/ProvinceDefinitions.h"
+#include "../../HOI4World/ProvinceDefinitions.h"
 #include "../../Mappers/Provinces/ProvinceMapper.h"
 #include "Log.h"
