+++ Vic2ToHoI4/Vic2ToHoI4.vcxproj.filters
@@ -178,9 +178,6 @@
     <ClCompile Include="..\common_items\CardinalToOrdinal.cpp">
       <Filter>CommonItems</Filter>
     </ClCompile>
-    <ClCompile Include="Source\Mappers\ProvinceDefinitions.cpp">
-      <Filter>Mappers</Filter>
-    </ClCompile>
     <ClCompile Include="Source\Mappers\GraphicsMapper.cpp">
       <Filter>Mappers</Filter>
     </ClCompile>
@@ -761,6 +758,9 @@
     <ClCompile Include="Source\V2World\Vic2Localisations.cpp">
       <Filter>Vic2World</Filter>
     </ClCompile>
+    <ClCompile Include="Source\HOI4World\ProvinceDefinitions.cpp">
+      <Filter>HoI4World</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="..\common_items\Date.h">
@@ -801,9 +801,6 @@
     </ClInclude>
     <ClInclude Include="..\common_items\CardinalToOrdinal.h">
       <Filter>CommonItems</Filter>
-    </ClInclude>
-    <ClInclude Include="Source\Mappers\ProvinceDefinitions.h">
-      <Filter>Mappers</Filter>
     </ClInclude>
     <ClInclude Include="Source\Mappers\GraphicsMapper.h">
       <Filter>Mappers</Filter>
@@ -1441,6 +1438,9 @@
     </ClInclude>
     <ClInclude Include="Source\V2World\Vic2Localisations.h">
       <Filter>Vic2World</Filter>
+    </ClInclude>
+    <ClInclude Include="Source\HOI4World\ProvinceDefinitions.h">
+      <Filter>HoI4World</Filter>
     </ClInclude>
   </ItemGroup>
   <ItemGroup>
