+++ Vic2ToHoI4Tests/Vic2ToHoI4Tests-Google.vcxproj.filters
@@ -136,9 +136,6 @@
     <ClCompile Include="..\Vic2ToHoI4\Source\Mappers\Provinces\VersionedMappings.cpp">
       <Filter>Vic2ToHoI4 files\Mappers\Provinces</Filter>
     </ClCompile>
-    <ClCompile Include="..\Vic2ToHoI4\Source\Mappers\ProvinceDefinitions.cpp">
-      <Filter>Vic2ToHoI4 files\Mappers</Filter>
-    </ClCompile>
     <ClCompile Include="MapperTests\TechMapperTests.cpp">
       <Filter>MapperTests</Filter>
     </ClCompile>
@@ -868,6 +865,9 @@
     </ClCompile>
     <ClCompile Include="..\Vic2ToHoI4\Source\V2World\Vic2Localisations.cpp">
       <Filter>Vic2ToHoI4 files\Vic2</Filter>
+    </ClCompile>
+    <ClCompile Include="..\Vic2ToHoI4\Source\HOI4World\ProvinceDefinitions.cpp">
+      <Filter>Vic2ToHoI4 files\HoI4</Filter>
     </ClCompile>
   </ItemGroup>
   <ItemGroup>
